use java based test
